Move roles files to separate folders. Update imports
