+++ src/app.module.ts
@@ -4,7 +4,7 @@ import { TypeOrmModule } from "@nestjs/typeorm";
 import { ConfigModule } from "@nestjs/config";
 import { UsersModule } from "./users/users.module";
 
-import { RolesGuard } from "./auth/roles/roles.guard";
+import { RolesGuard } from "./auth/guards/roles.guard";
 
 import { AuthModule } from "./auth/auth.module";
 import { ProductsModule } from "./products/products.module";
